requirements.txt
PySide6>=6.6,<7.0
reportlab>=4.0
qrcode[pil]>=7.4
Pillow>=10.0
pyinstaller>=6.0
